refactor: ensure synchronous form reset gating and extract error-scrolling hook (#345)
- Introduced isContainerLoading state to ensure form.reset() runs in useLayoutEffect synchronously before the form tree mounts, protecting cascading dropdown lifecycles.
- Created highly reusable useFormErrorScroll hook to abstract coordinates parsing, sorting, and formRef scoping.
- Added container.test.tsx integration test to verify that form.reset gets correctly triggered with mapped values.
- Cleaned up all unused imports and variables in container.tsx.
